Updated UI
git-svn-id: file:///fltk/svn/branches/branch-3.0@8971 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ fluid/workspace_panel.cxx
@@ -87,7 +87,7 @@ pEnvMenu->copy(menu_pEnvMenu);
 fltk3::MenuItem *mi = (fltk3::MenuItem*)pEnvMenu->menu();
 int level = 0;
 while (mi->label()) {
-  unsigned int me = (intptr_t)mi->user_data();
+  unsigned int me = (int)(long)mi->user_data();
   if (me) {
     fltk3::MultiLabel *ml = new fltk3::MultiLabel;
     ml->labela = (char*)&menu_none_pixmap;
